centralize sort logic
